+++ NEWS.md
@@ -1,3 +1,3 @@
-# microtiterr 0.1.0
+# microtiterr 0.1.0 - 2017-04-04
 
-* Added a `NEWS.md` file to track changes to the package.
+* Initial release with functions `is.well`, `microtiter_data_frame`, `microtiter_tibble`, `row_letters`, `well_column`, `well_label`, and `well_row`